requirements.txt
playwright
pytest
requests
PyYAML
python-dotenv
faker
jsonschema
